ORM: Fixed a couple of bugs and ManyMany now tested to make and break relationships.

+++ fuel/packages/orm/classes/hasone.php
@@ -169,7 +169,7 @@ class HasOne extends Relation {
 		}
 
 		$cascade = is_null($cascade) ? $this->cascade_save : (bool) $cascade;
-		if ($cascade)
+		if ($cascade and ! empty($model_to))
 		{
 			$model_to->delete();
 		}
